chore: move extension to a separate folder
