+++ lib/src/helpers/extensions.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/cupertino.dart';
+
+extension BuildContextExt on BuildContext {
+  double get screenWidth => MediaQuery.of(this).size.width;
+}
