+++ lib/src/models/pager_indicator_view_model.dart
@@ -1,4 +1,4 @@
-import 'package:intro_views_flutter/src/constants/constants.dart';
+import 'package:intro_views_flutter/src/helpers/constants.dart';
 import 'package:intro_views_flutter/src/models/page_view_model.dart';
 
 /// View model for page indicator.
