+++ lib/src/ui/intro_views_flutter.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:intro_views_flutter/src/animation_gesture/animated_page_dragger.dart';
 import 'package:intro_views_flutter/src/animation_gesture/page_dragger.dart';
 import 'package:intro_views_flutter/src/animation_gesture/page_reveal.dart';
-import 'package:intro_views_flutter/src/constants/constants.dart';
+import 'package:intro_views_flutter/src/helpers/constants.dart';
 import 'package:intro_views_flutter/src/models/page_view_model.dart';
 import 'package:intro_views_flutter/src/models/pager_indicator_view_model.dart';
 import 'package:intro_views_flutter/src/models/slide_update_model.dart';
